개행, 접근제한자 및 기타 수정
